drilling down into grids to understand the affects to each item on the user page.
