+++ src/components/AdminPage/AdminPage.js
@@ -7,15 +7,19 @@ import { triggerLogout } from '../../redux/actions/loginActions';
 import AdminMethods from '../AdminMethods/AdminMethods';
 import NewMethodDialog from '../NewMethodDialog/NewMethodDialog';
 import Button from 'material-ui/Button';
+import Grid from 'material-ui/Grid'
 
 const styles = {
-  root: {
+  root: {       
+      flexGrow: 1,
       display: 'flex',
       flexWrap: 'wrap',
       justifyContent: 'space-around',
-      maxWidth: '960px'        
+      justify: 'center',
+      alignItems: 'center',
+      maxWidth: '960px',
   }
-};
+}
 
 
 const mapStateToProps = state => ({
@@ -43,23 +47,23 @@ class AdminPage extends Component {
 
     if (this.props.user.userName) {
       content = (
-        <div>
-          <div>
+        <div style={styles.root}>
+          <Grid item xs={12}>
             <h1 id="welcome">
               Welcome, { this.props.user.userName }!
             </h1>
             <Button onClick={this.logout}>
               Log Out
             </Button>
-          </div>
-          <div>
+          </Grid>
+          <Grid item xs={12}>
             <p>Long paragraph of explaination on how to use this stuff in the admin's page.</p>
             <NewMethodDialog />
-          </div>
-          <div>
+          </Grid>
+          <Grid item xs={12}>
             List of methods
             <AdminMethods />
-          </div>
+          </Grid>
         </div>
       );
     }
