+++ src/components/UserMethods/UserMethods.js
@@ -4,9 +4,8 @@ import UserMethodItem from '../UserMethodItem/UserMethodItem';
 
 const styles = {
     root: {
-        display: 'flex',
+        justifyContent: 'space-around',
         flexWrap: 'wrap',
-        maxWidth: '960px'
     }
 };
